Cluster creation and glusterd start in framework
The handling of cluster start and glusterd start will
be done at the main level. If it is nonDisruptive TCs,
then the cluster or glusterd working won't be tampered with
unless and until there's some serious issue or a new bug.

It is only for the disruptive cases wherein the cluster recreation
and glusterd startup will come handy.

Fixes: #220

Signed-off-by: srijan-sivakumar <ssivakum@redhat.com>