Converting threading to multiprocessing (#230)
* Converting threading to multiprocessing

The current implementation will have a bottleneck because of
the nature of how threads are executed in python, wherein threads
and the main process share the GIL even though the system might
boast of multiple processor cores.

To work around this bottleneck in python, multiprocessing module
can be used. Now one can argue that multiprocess will create a
separate process in the background and that this will be heavier than
a thread and also would consume more time than spinning up a new thread.

To solve this, we use the idea of worker processes taking up tasks from
a queue, hence processes are spun up once and are killed only when there
is no job in the queue.

One more point to note is that the multiprocessing requires some mechanism of
locking to prevent the main process form going ahead with its flow ( which in
our case would be the execution of the disruptive test cases ). Now for the time
being we have an incremental backoff system with sleep till all jobs are dead.

This will have to be replaced with a sophesticated locking and signalling mechanism.

Fixes: #229

Signed-off-by: srijan-sivakumar <ssivakum@redhat.com>

* Test result handler working now

The issue was the Queue was not updating in case of non_concurrent tests
which gets handled directly in the run_test now and for the concurrent
tests the volume types are aggregated as well providing a single table
for all the volumes for a test.

Signed-off-by: Ayush Ujjwal <aujjwal@redhat.com>

* removed the extra print

Signed-off-by: Ayush Ujjwal <aujjwal@redhat.com>

* Modified the code and refactored the result aggregation code

Signed-off-by: Ayush Ujjwal <aujjwal@redhat.com>

* Using the queue only to get the work done

Signed-off-by: Ayush Ujjwal <aujjwal@redhat.com>

* Update test_runner.py

* Update test_runner.py

Co-authored-by: Ayush Ujjwal <aujjwal@redhat.com>